restructure frontend set-up
