+++ frontend/checkpoint-04-04-25/src/App.css
@@ -1,0 +1,13 @@
+@import "tailwindcss";
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+}
+
+#table-container {
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+}
