+++ frontend/checkpoint-04-04-25/src/pages/Inventory.jsx
